test: tag some tests
