+++ .github/workflows/ci-core-integration.yml
@@ -1,0 +1,136 @@
+name: CI Core Integration
+run-name: CI Core Integration
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}-${{ github.event_name }}
+  cancel-in-progress: true
+
+# Run on key branches to make sure integration is good, otherwise run on all PR's
+on:
+  push:
+    branches:
+      - develop
+      - "release/*"
+  merge_group:
+  pull_request:
+
+
+jobs:
+  core:
+    env:
+      # We explicitly have this env var not be "CL_DATABASE_URL" to avoid having it be used by core related tests
+      # when they should not be using it, while still allowing us to DRY up the setup
+      DB_URL: postgresql://postgres:postgres@localhost:5432/chainlink_test?sslmode=disable
+    strategy:
+      fail-fast: false
+      matrix:
+        type:
+          - module: Deployment
+            cmd: deployment_tests
+            os: runs-on=${{ github.run_id }}-deployment/cpu=48/ram=96/family=c6id/spot=false/image=ubuntu24-full-x64/extras=s3-cache
+            tags: "integration solana"
+            go-mod-directory: "deployment/"
+
+    name: "${{ matrix.type.module }} - Tagged Tests (${{matrix.type.tags}})"
+    timeout-minutes: 20
+    runs-on: ${{ matrix.type.os }}
+    permissions:
+      id-token: write
+      contents: read
+      actions: read
+    steps:
+      - name: Enable S3 Cache for Self-Hosted Runners
+        # these env vars are set (and exposed) when it is a self-hosted runner with extras=s3-cache
+        if: ${{ env.RUNS_ON_INSTANCE_ID != '' && env.ACTIONS_CACHE_URL != '' }}
+        uses: runs-on/action@66d4449b717b5462159659523d1241051ff470b9 # v1
+
+      - name: Checkout the repo
+        uses: actions/checkout@v4
+        with:
+          persist-credentials: false
+
+      - name: Setup Go
+        uses: ./.github/actions/setup-go
+        with:
+          # race/fuzz tests don't benefit repeated caching, so restore from develop's build cache
+          restore-build-cache-only: ${{ matrix.type.cmd == 'go_core_fuzz' || matrix.type.cmd == 'go_core_race_tests' }}
+          build-cache-version: ${{ matrix.type.cmd }}
+          go-version-file: ${{ matrix.type.go-mod-directory }}go.mod
+          go-module-file: ${{ matrix.type.go-mod-directory }}go.sum
+
+      - name: Setup Solana
+        uses: ./.github/actions/setup-solana
+
+      # - name: Setup Aptos
+      #   if: ${{ matrix.type.should-run == 'true' && matrix.type.setup-aptos == 'true' }}
+      #   uses: aptos-labs/actions/install-aptos-cli@63740b290d839b87ecfafbcf75ed03a36a54a29f # jan 15, 2025
+      #   with:
+      #     CLI_VERSION: 7.10.0
+
+      # - name: Setup Sui CLI v1.57.2
+      #   if: ${{ matrix.type.should-run == 'true' && matrix.type.setup-sui == 'true' }}
+      #   uses: ./.github/actions/setup-sui
+      #   with:
+      #     github-token: ${{ secrets.GITHUB_TOKEN }}
+      #     version: mainnet-1.57.2
+
+      # - name: Setup wasmd
+      #   if: ${{ matrix.type.should-run == 'true' }}
+      #   uses: ./.github/actions/setup-wasmd
+
+      - name: Setup Postgres
+        if: ${{ matrix.type.should-run == 'true' }}
+        uses: smartcontractkit/.github/actions/setup-postgres@setup-postgres/v1
+        with:
+          tmpfs: "true"
+          image-tag: "16-alpine"
+
+      - name: Touching core/web/assets/index.html
+        if: ${{ matrix.type.should-run == 'true' }}
+        run: mkdir -p core/web/assets && touch core/web/assets/index.html
+
+      - name: Download Go vendor packages
+        if: ${{ matrix.type.should-run == 'true' }}
+        run: go mod download
+
+      - name: Setup DB
+        if: ${{ matrix.type.should-run == 'true' }}
+        run: go run ./core/store/cmd/preparetest
+        env:
+          CL_DATABASE_URL: ${{ env.DB_URL }}
+
+      # - name: Install LOOP Plugins
+      #   if: ${{ matrix.type.should-run == 'true' && matrix.type.install-loopps == 'true' }}
+      #   run: make install-plugins
+
+      - name: Run tests
+        if: ${{ matrix.type.should-run == 'true' }}
+        timeout-minutes: ${{ github.event_name == 'schedule' && 37 || 22 }} # leave time for remaining steps
+        continue-on-error: ${{ matrix.type.trunk-auto-quarantine == 'true' }}
+        id: run-tests
+        shell: bash
+        env:
+          GO_TEST_TAGS: ${{ matrix.type.tags }}
+          CL_DATABASE_URL: ${{ env.DB_URL }}
+          OUTPUT_FILE: ./output.txt
+          PRODUCE_JUNIT_XML: ${{ matrix.type.trunk-auto-quarantine }}
+          JUNIT_FILE: ${{ github.workspace }}/junit.xml
+          RUN_QUARANTINED_TESTS: "true" # always run quarantined tests in CI
+        run: |
+          GODEBUG=goindex=0 ./tools/bin/${{ matrix.type.cmd }} ./...
+          # See: https://github.com/golang/go/issues/69179
+
+      - name: Store logs artifacts
+        if: ${{ always() && matrix.type.should-run == 'true' }}
+        uses: actions/upload-artifact@v4
+        with:
+          name: ${{ matrix.type.cmd }}_logs
+          path: |
+            ./output.txt
+            ./output-short.txt
+            ./race.*
+            ./coverage.txt
+            ./postgres_logs.txt
+            ./junit.xml
+            ./deployment/junit.xml
+          retention-days: 7
