+++ deployment/ccip/changeset/solana_v0_1_1/testutils_test.go
@@ -1,0 +1,13 @@
+package solana_test
+
+import (
+	"os"
+	"testing"
+)
+
+func skipInCI(t *testing.T) {
+	ci := os.Getenv("CI") == "true"
+	if ci && false { // test - disable
+		t.Skip("Skipping in CI")
+	}
+}
